+++ src/js/api/global.js
@@ -1,4 +1,4 @@
-import {$, apply, isString, mergeOptions, parents, toNode} from 'uikit-util';
+import {$, apply, fastdom, isString, mergeOptions, parents, toNode} from 'uikit-util';
 
 export default function (UIkit) {
 
@@ -75,5 +75,4 @@ export default function (UIkit) {
         }
 
     }
-
 }