+++ src/js/core/drop.js
@@ -1,6 +1,6 @@
 import Position from '../mixin/position';
 import Togglable from '../mixin/togglable';
-import {addClass, Animation, attr, children, css, includes, isTouch, matches, MouseTracker, offset, on, once, pointerEnter, pointerLeave, query, removeClasses, toggleClass, trigger, within} from 'uikit-util';
+import {addClass, Animation, apply, attr, css, includes, isTouch, matches, MouseTracker, offset, on, once, pointerEnter, pointerLeave, query, removeClasses, toggleClass, trigger, within} from 'uikit-util';
 
 let active;
 
@@ -384,8 +384,9 @@ export default {
 };
 
 function getPositionedElements(el) {
-    const result = css(el, 'position') !== 'static' ? [el] : [];
-    return result.concat.apply(result, children(el).map(getPositionedElements));
+    const result = [];
+    apply(el, el => css(el, 'position') !== 'static' && result.push(el));
+    return result;
 }
 
 export function delayOn(el, type, fn) {